requirements.txt
aiohttp>=3.7.4.post0
async-timeout>=3.0.1
attrs>=21.2.0
beautifulsoup4>=4.10.0
certifi>=2021.10.8
cffi>=1.15.0
chardet>=4.0.0
charset-normalizer>=2.0.7
discord.py>=1.7.3
DiscordUtils>=1.3.4
idna>=3.3
lyricsgenius>=3.0.1
multidict>=5.2.0
pycparser>=2.20
PyNaCl>=1.4.0
requests>=2.26.0
six>=1.16.0
soupsieve>=2.2.1
typing-extensions>=3.10.0.2
urllib3>=1.26.7
yarl>=1.7.0
yt-dlp
